修复Long64写入错误的int类型

+++ src/main/java/io/github/yezhihao/protostar/schema/NumberSchema.java
@@ -111,7 +111,7 @@ public class NumberSchema {
 
         @Override
         public void writeTo(ByteBuf output, Long value) {
-            output.writeLong(value.intValue());
+            output.writeLong(value);
         }
     }
 }
